requirements.txt
configparser==3.7.4
pickleDB==0.9.2
python-telegram-bot==11.1.0
black==19.3b0
